+++ packages/ui/src/core/modal/modal.component.tsx
@@ -15,7 +15,10 @@ const isValidSingleFragmentChildren = (children?: any): boolean =>
   (children as JSX.Element).type === React.Fragment;
 
 const Modal = intrinsicComponent<ModalProps, HTMLDivElement>(
-  ({ children: _children, open, onClose, maxWidth, fullWidth, modalStyles, ...rest }, ref): JSX.Element => {
+  (
+    { children: _children, open, onClose, maxWidth, fullWidth, modalStyles, disableOverlayClick, ...rest },
+    ref
+  ): JSX.Element => {
     const children = isValidSingleFragmentChildren(_children) ? (_children as JSX.Element).props.children : _children;
 
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
@@ -52,7 +55,7 @@ const Modal = intrinsicComponent<ModalProps, HTMLDivElement>(
     const render = (): JSX.Element => (
       <ModalMenuContext.Provider value={{ modalOpened: Boolean(open) }}>
         <Styled.Wrapper style={{ ...modalStyles }} open={Boolean(open)} ref={ref}>
-          <Styled.Overlay onClick={handleClose} open={Boolean(open)} />
+          <Styled.Overlay onClick={() => (disableOverlayClick ? null : handleClose())} open={Boolean(open)} />
 
           <Styled.Container {...rest} maxWidth={maxWidth} fullWidth={fullWidth} open={Boolean(open)}>
             <Styled.Modal>
@@ -79,6 +82,7 @@ const Modal = intrinsicComponent<ModalProps, HTMLDivElement>(
 export const defaultProps = {
   open: false,
   fullWidth: false,
+  disableOverlayClick: false,
   maxWidth: Size.Xs,
 };
 
@@ -87,6 +91,7 @@ Modal.defaultProps = defaultProps;
 export const propTypes = {
   onClose: PT.func,
   modalStyles: PT.object,
+  disableOverlayClick: PT.bool,
   children: PT.node.isRequired,
   maxWidth: PT.oneOf(objectValues(Size)),
   open: PT.bool,