+++ packages/ui/src/core/popup-status/popup-status.component.tsx
@@ -12,19 +12,25 @@ import Styled from './popup-status.styles';
 const getPopupStatusIcon = (status?: PopupStatusType, notificationBackground?: boolean): JSX.Element => {
   switch (status) {
     case Status.Warning:
-      return notificationBackground ? 
-      <Warning color={lightPalette[Color.Warning]} size={20} />
-      : <WarningOutline color={lightPalette[Color.Warning]} size={12} />;
+      return notificationBackground ? (
+        <Warning color={lightPalette[Color.Warning]} size={20} />
+      ) : (
+        <WarningOutline color={lightPalette[Color.Warning]} size={12} />
+      );
 
     case Status.Error:
-      return notificationBackground ?
-      <Error color={lightPalette[Color.Error]} size={20} />
-      : <ErrorOutline color={lightPalette[Color.Error]} size={12} />;
+      return notificationBackground ? (
+        <Error color={lightPalette[Color.Error]} size={20} />
+      ) : (
+        <ErrorOutline color={lightPalette[Color.Error]} size={12} />
+      );
 
     case Status.Info:
-      return notificationBackground ?
-      <About color={lightPalette[Color.Info]} size={20} />
-      : <InfoOutline color={lightPalette[Color.Info]} size={12} />;
+      return notificationBackground ? (
+        <About color={lightPalette[Color.Info]} size={20} />
+      ) : (
+        <InfoOutline color={lightPalette[Color.Info]} size={12} />
+      );
 
     case Status.Success:
     default: