+++ packages/ui/src/core/tag-field/tag-field.props.d.ts
@@ -43,11 +43,15 @@ export interface TagFieldProps extends React.HTMLAttributes<HTMLDivElement> {
   clearTagsButtonLabel?: string;
   hideCopyIcon?: boolean;
   alwaysShowSuggestedTags?: boolean;
+  placeholderAlwaysVisible?: boolean;
   suggestionsFilter?: SuggestionsFilterFnType;
   suggestionLabel?: React.ReactNode;
   suggestionTooltipMessage?: string;
   copyTextMessage?: string;
   copySuccessIcon?: React.ReactNode | ((props: IconProps) => React.ReactNode);
   onFocus?: React.FocusEventHandler<HTMLInputElement>;
-  onBlur?: React.FocusEventHandler<HTMLInputElement>;
+  onBlur?: (event: React.FocusEvent<HTMLInputElement>, setUserInput: (value: any) => void) => void;
+  filterInputWidth?: string | number;
+  fullWidth?: boolean;
+  hideClearButton?: boolean;
 }