+++ packages/ui/src/core/textarea/textarea.styles.ts
@@ -38,6 +38,7 @@ const Textarea = styled.div.attrs({
     color: ${disabled ? theme.palette[PColor.TextPlaceholder] : theme.palette[PColor.TextPrimary]};
 
     ${heightTextAreaMixin[size]}
+
     ${!readOnly &&
     !disabled &&
     css`
@@ -73,7 +74,7 @@ const Textarea = styled.div.attrs({
 const Base = styled.textarea.attrs({
   className: generateClassNames(baseClassName, 'Base'),
 })<TextareaProps>(
-  ({ theme: { palette }, size = InputSize.Md, readOnly = false }: With<WithTheme, TextareaProps>) => css`
+  ({ theme: { palette }, readOnly = false, size = InputSize.Md }: With<WithTheme, TextareaProps>) => css`
     display: block;
     width: 100%;
     height: 100%;
@@ -92,10 +93,10 @@ const Base = styled.textarea.attrs({
     font-weight: inherit;
     font-family: inherit;
     box-sizing: border-box;
-
-    ${sizeTextAreaMixin[size]};
+    flex: 1;
 
     ${readOnly && `padding-bottom: 0px;`}
+    ${sizeTextAreaMixin[size]};
 
     &::placeholder {
       color: ${palette[PColor.TextPlaceholder]};
@@ -117,29 +118,29 @@ const Base = styled.textarea.attrs({
   `
 );
 
-const CopyIcon = styled.div.attrs({
-  className: generateClassNames(baseClassName, 'CopyIcon'),
+const ActionsButtonsWrapper = styled.div.attrs({
+  className: generateClassNames(baseClassName, 'ActionsButtonsWrapper'),
 })(
-  ({
-    size = InputSize.Md,
-    isHovering = false,
-    theme: { palette },
-  }: With<With<Partial<TextareaProps>, { isHovering: boolean }>, WithTheme>) => css`
+  ({ size = InputSize.Md }: Partial<TextareaProps>) => css`
     display: flex;
-    justify-content: flex-end;
+    align-items: center;
     width: 100%;
-    cursor: pointer;
-    color: ${palette[PColor.IconsPrimary]};
+    column-gap: 12px;
+    box-sizing: border-box;
     ${sizeTextAreaMixin[size]};
-    ${!isHovering &&
-    `
-      svg {
-        visibility: hidden;
-      }
-    `}
   `
 );
 
-const Styled = applyDisplayNames({ Textarea, CopyIcon, Base });
+const CopyIcon = styled.span.attrs({
+  className: generateClassNames(baseClassName, 'CopyIcon'),
+})(
+  ({ theme: { palette } }: With<With<Partial<TextareaProps>, { showCopyIcon: boolean }>, WithTheme>) => css`
+    cursor: pointer;
+    color: ${palette[PColor.IconsPrimary]};
+    margin-left: auto;
+  `
+);
+
+const Styled = applyDisplayNames({ Textarea, CopyIcon, Base, ActionsButtonsWrapper });
 
 export default Styled;