+++ packages/ui/stories/core/tag-field.stories.tsx
@@ -35,8 +35,8 @@ const defaultArgs = {
   hideCopyIcon: false,
   showGenerateTagsButton: true,
   suggestionTooltipMessage: 'Suggested tags',
-  copyTextMessage: 'Link copied to clipboard',
   copySuccessIcon: <Success size={16} />,
+  copyTextMessage: 'Tags copied to clipboard',
 };
 
 const handleRemove = (