+++ packages/ui/stories/core/textarea.stories.tsx
@@ -19,6 +19,13 @@ const defaultArgs = {
   placeholder: 'placeholder',
   readOnly: false,
   disabled: false,
+  showActionButton: false,
+  showClearButton: false,
+  showCopyIcon: false,
+  disableActionButton: false,
+  isActionButtonLoading: false,
+  actionButtonLabel: 'Action',
+  clearAllButtonLabel: 'Clear all',
   copyTextMessage: 'Link copied to clipboard',
   copySuccessIcon: <Success size={16} />,
 };